feat: AuthGate 컴포넌트 구현 - 접근성 관리 및 로그인 유지, 프로젝트 작성 버튼 미로그인시 비활성화
