+++ src/app/api/user/login/route.tsx
@@ -18,8 +18,8 @@ export async function POST(request: NextRequest) {
     // TODO: 실제 로그인 로직 구현 필요
     // 예시 응답
     const mockResponse = {
-      "accessToken": userJWT,
-      "refreshToken": userJWT,
+      "accessToken": adminJWT,
+      "refreshToken": adminJWT,
       "userId": 9007199254740991
     };
