+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import localFont from "next/font/local"; // 지정한 폰트를 사용하기 위
 import "./globals.css";
 import HeaderLoader from "@/components/HeaderLoader";
 import RegisterPageButtonLoader from "@/components/RegisterPageButtonLoader";
-import { AdminGuard } from "@/components/login/AdminGuard";
+import { AuthGate } from "@/components/login/AuthGate";
 
 // Pretendard 폰트를 사용하기 위해 불러옴
 const pretendard = localFont({
@@ -28,7 +28,7 @@ export default function RootLayout({
       <body
         className={`${pretendard.variable} font-pretendard`} // Pretendard 사용
       >
-        <AdminGuard />
+        <AuthGate />
         <HeaderLoader />
         <main className="flex-1 bg-white mt-1">
           <div className="container mx-auto px-4 py-8">
