+++ src/components/projectRegister/RegisterPageButton.tsx
@@ -1,9 +1,14 @@
 'use client'
 
 import { useRouter } from "next/navigation";
-
+import useAuthStore from "@/stores/auth";
 export default function RegisterPageButton() {
+    const isLoggedIn = useAuthStore((s) => s.isLoggedIn);
     const router = useRouter();
+
+    if (!isLoggedIn) {
+        return null;
+    }
 
     return (
         <button
